code quality: don't use uint64_t when dealing with memory addresses but uintptr_t.

+++ src/hooks/call_hook.cpp
@@ -39,7 +39,7 @@ namespace big
 	{
 		auto seq                       = g_call_hook_memory.allocate_jump_sequence(hook);
 		m_patched_bytes[0]             = 0xE8;
-		*(int32_t*)&m_patched_bytes[1] = (int32_t)((uint64_t)seq - (uint64_t)location - 5);
+		*(int32_t*)&m_patched_bytes[1] = (int32_t)((uintptr_t)seq - (uintptr_t)location - 5);
 		memcpy(m_original_bytes, location, 5);
 		m_original_function = memory::handle(location).add(1).rip().as<void*>();
 	}
