+++ src/logger/exception_handler.cpp
@@ -9,10 +9,10 @@
 
 namespace big
 {
-	static auto hash_stack_trace(std::vector<uint64_t> stack_trace)
+	static auto hash_stack_trace(std::vector<uintptr_t> stack_trace)
 	{
 		auto data        = reinterpret_cast<const char*>(stack_trace.data());
-		std::size_t size = stack_trace.size() * sizeof(uint64_t);
+		std::size_t size = stack_trace.size() * sizeof(uintptr_t);
 
 		return std::hash<std::string_view>()({data, size});
 	}
